[ #5670 ] Added the module Agda.Syntax.Scope.Flat.
And moved FlatScope, flattenScope, getDefinedNames and localNames to
that module.
