+++ src/full/Agda/Syntax/Concrete/Operators.hs
@@ -19,7 +19,6 @@ module Agda.Syntax.Concrete.Operators
     ) where
 
 import Control.Applicative ( Alternative((<|>)))
-import Control.Arrow (second)
 import Control.Monad.Except (throwError)
 
 import Data.Either (partitionEithers)
@@ -42,6 +41,7 @@ import qualified Agda.Syntax.Abstract.Name as A
 import Agda.Syntax.Position
 import Agda.Syntax.Notation
 import Agda.Syntax.Scope.Base
+import Agda.Syntax.Scope.Flat
 import Agda.Syntax.Scope.Monad
 
 import Agda.TypeChecking.Monad.Base (typeError, TypeError(..), LHSOrPatSyn(..))
@@ -80,45 +80,6 @@ billToParser k = Bench.billTo
 ---------------------------------------------------------------------------
 -- * Building the parser
 ---------------------------------------------------------------------------
-
-type FlatScope = Map QName [AbstractName]
-
--- | Compute all defined names in scope and their fixities/notations.
--- Note that overloaded names (constructors) can have several
--- fixities/notations. Then we 'mergeNotations'. (See issue 1194.)
-getDefinedNames :: KindsOfNames -> FlatScope -> [[NewNotation]]
-getDefinedNames kinds names =
-  [ mergeNotations $ map (namesToNotation x . A.qnameName . anameName) ds
-  | (x, ds) <- Map.toList names
-  , any ((`elemKindsOfNames` kinds) . anameKind) ds
-  , not (null ds)
-  ]
-  -- Andreas, 2013-03-21 see Issue 822
-  -- Names can have different kinds, i.e., 'defined' and 'constructor'.
-  -- We need to consider all names that have *any* matching kind,
-  -- not only those whose first appearing kind is matching.
-
--- | Compute all names (first component) and operators/notations
--- (second component) in scope.
-localNames :: FlatScope -> ScopeM ([QName], [NewNotation])
-localNames flat = do
-  let defs = getDefinedNames allKindsOfNames flat
-  locals <- nubOn fst . notShadowedLocals <$> getLocalVars
-  -- Note: Debug printout aligned with the one in buildParsers.
-  reportS "scope.operators" 50
-    [ "flat  = " ++ prettyShow flat
-    , "defs  = " ++ prettyShow defs
-    , "locals= " ++ prettyShow locals
-    ]
-  let localNots  = map localOp locals
-      notLocal   = not . hasElem (map notaName localNots) . notaName
-      otherNots  = concatMap (filter notLocal) defs
-  return $ second (map useDefaultFixity) $ split $ localNots ++ otherNots
-  where
-    localOp (x, y) = namesToNotation (QName x) y
-    split          = partitionEithers . concatMap opOrNot
-    opOrNot n      = Left (notaName n) :
-                     [Right n | not (null (notation n))]
 
 -- | A data structure used internally by 'buildParsers'.
 data InternalParsers e = InternalParsers
